added option to not be destructive

+++ crates/cli/src/main.rs
@@ -3,10 +3,7 @@
 use std::{fs::File, io::Write, path::PathBuf};
 
 use genlib::{
-    generators::{
-        python::FastApi,
-        ts::{self, TS},
-    },
+    generators::{python::FastApi, ts::TS},
     parser::definitions::*,
 };
 
@@ -17,6 +14,9 @@ use clap::{self, Arg, ArgAction, Args, Parser, Subcommand, arg};
 #[derive(Parser)]
 struct Cli {
     pub definitions: Vec<PathBuf>,
+
+    #[arg(short, long, default_value_t = false)]
+    pub desctructive: bool,
 
     #[arg(short, long, default_value_t = false)]
     pub verbose: bool,
@@ -87,15 +87,19 @@ fn main() -> Result<()> {
             type_path.push(format!("{prefix}types_{name}{postfix}"));
             type_path.set_extension(extension);
 
-            let mut type_file = File::create(type_path)?;
-            type_file.write_all(type_code.as_bytes())?;
+            if cli.desctructive || (!cli.desctructive && !type_path.exists()) {
+                let mut type_file = File::create(type_path)?;
+                type_file.write_all(type_code.as_bytes())?;
+            }
 
             let mut endpoint_path = cli.path.clone();
             endpoint_path.push(format!("{prefix}endpoint_{name}{postfix}"));
             endpoint_path.set_extension(extension);
 
-            let mut endpoint_file = File::create(endpoint_path)?;
-            endpoint_file.write_all(endpoint_code.as_bytes())?;
+            if cli.desctructive || (!cli.desctructive && !endpoint_path.exists()) {
+                let mut endpoint_file = File::create(endpoint_path)?;
+                endpoint_file.write_all(endpoint_code.as_bytes())?;
+            }
         } else {
             if cli.verbose {
                 println!("generated united joined...");
@@ -108,8 +112,10 @@ fn main() -> Result<()> {
             path.push(format!("{prefix}{name}{postfix}"));
             path.set_extension(extension);
 
-            let mut file = File::create(path)?;
-            file.write_all(code.as_bytes())?;
+            if cli.desctructive || (!cli.desctructive && !path.exists()) {
+                let mut file = File::create(path)?;
+                file.write_all(code.as_bytes())?;
+            }
         }
     } else {
         if cli.split {
@@ -122,8 +128,10 @@ fn main() -> Result<()> {
                 type_path.push(format!("{prefix}types_{name}{postfix}"));
                 type_path.set_extension(extension);
 
-                let mut type_file = File::create(type_path)?;
-                type_file.write_all(type_code.as_bytes())?;
+                if cli.desctructive || (!cli.desctructive && !type_path.exists()) {
+                    let mut type_file = File::create(type_path)?;
+                    type_file.write_all(type_code.as_bytes())?;
+                }
             }
 
             for (name, endpoint_code) in defs.build_decoupled_endpoint_module(&*generator) {
@@ -132,8 +140,10 @@ fn main() -> Result<()> {
                 endpoint_path.push(format!("{prefix}endpoints_{name}{postfix}"));
                 endpoint_path.set_extension(extension);
 
-                let mut endpoint_file = File::create(endpoint_path)?;
-                endpoint_file.write_all(endpoint_code.as_bytes())?;
+                if cli.desctructive || (!cli.desctructive && !endpoint_path.exists()) {
+                    let mut endpoint_file = File::create(endpoint_path)?;
+                    endpoint_file.write_all(endpoint_code.as_bytes())?;
+                }
             }
         } else {
             if cli.verbose {
@@ -145,8 +155,10 @@ fn main() -> Result<()> {
                 path.push(format!("{prefix}{name}{postfix}"));
                 path.set_extension(extension);
 
-                let mut file = File::create(path)?;
-                file.write_all(code.as_bytes())?;
+                if cli.desctructive || (!cli.desctructive && path.exists()) {
+                    let mut file = File::create(path)?;
+                    file.write_all(code.as_bytes())?;
+                }
             }
         }
     }
